pyproject.toml
[project]
name = "taqc"
version = "0.4.0"
description = ""
authors = [{ name = "Day Tripper", email = "kekovdom@gmail.com" }]
dependencies = [
    "more-itertools>=10.2.0",
    "pillow<10.3.0",
    "pytest>=8.1.1",
    "expression>=4.3.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pdm]
distribution = true


[tool.pyright]
executionEnvironments = [{ root = "." }]
venvPath = "."
venv = ".venv"
